+++ crates/byte-fmt/src/lib.rs
@@ -60,12 +60,13 @@ pub fn b64(b: &[u8], mini: bool) -> String {
 }
 pub use abe::scope::base::{base64,base64_decode,mini_b64};
 
+use bytemuck::{Pod,Zeroable};
 
-#[derive(Default, Copy, Clone, Hash, PartialEq, Eq, PartialOrd, Ord)]
+#[derive(Default, Copy, Clone, Hash, PartialEq, Eq, PartialOrd, Ord,Pod,Zeroable)]
 #[repr(transparent)]
 /// newtype around bytes to print/parse [[abe]] text
 pub struct AB<N: ?Sized = Vec<u8>>(pub N);
-#[derive(Default, Copy, Clone, Hash, PartialEq, Eq, PartialOrd, Ord)]
+#[derive(Default, Copy, Clone, Hash, PartialEq, Eq, PartialOrd, Ord,Pod,Zeroable)]
 #[repr(transparent)]
 /// newtype around bytes to print/parse b64 (url-safe no-padding)
 pub struct B64<N = [u8; 32]>(pub N);