+++ crates/pkt/src/netpkt/netpkt_arc.rs
@@ -3,7 +3,7 @@
 // This Source Code Form is subject to the terms of the Mozilla Public
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at https://mozilla.org/MPL/2.0/.
-use crate::{NetPkt, NetPktPtr, PartialNetHeader, Error};
+use crate::{NetPkt, NetPktPtr, PartialNetHeader, Error };
 use std::{
     borrow::Borrow, fmt::Debug, mem::size_of, ops::Deref, ptr::{ self}, sync::atomic::AtomicUsize,
 };
@@ -50,7 +50,7 @@ impl NetPktArcPtr {
     /// Get a look at the outer [NetPktArc] 
     #[inline]
     pub fn with_arc<U>(&self, f: impl FnOnce(&NetPktArc) -> U) -> U {
-        let size : usize = self.as_point().point_header_ref().content_size().into();
+        let size : usize = self.as_point().point_header_ref().padded_content_size().into();
         let ptr: *const NetPktPtr = ptr::from_ref(self.netpktptr());
         let inner_arc : *const () =unsafe{ ptr.cast::<u8>().sub(size_of::<AtomicUsize>()).cast::<()>()};
 
@@ -86,9 +86,9 @@ impl NetPktArc {
     /// must be created through [Self::into_raw_arc]
     pub unsafe fn from_raw_arc(ptr: *const NetPktArcPtr) -> Self {
         assert!(!ptr.is_null());
-        let size : usize = unsafe { (*ptr).as_point().point_header_ref().content_size() }.into();
+        let byte_size : usize = unsafe { (*ptr).as_point().point_header_ref().padded_content_size() }.into();
         let inner_arc: *const () = ptr.cast::<u8>().sub(size_of::<AtomicUsize>()).cast::<()>();
-        let inner: *const ArcInner = std::ptr::from_raw_parts(inner_arc, size);
+        let inner: *const ArcInner = std::ptr::from_raw_parts(inner_arc, byte_size);
         let arc: Arc<HeaderSlice<NetPktPtr, [u8]>> = unsafe { std::mem::transmute(inner) };
         NetPktArc(arc)
     }
@@ -101,8 +101,12 @@ impl NetPktArc {
             hash: partial.hash,
             point: crate::PointThinPtr(partial.point_header),
         };
-        let inner_len : usize= h.point.point_header().content_size().into() ;
-        let pkt= NetPktArc(triomphe::Arc::from_header_and_fn(h,inner_len,copy_from));
+        let byte_size : usize= h.point.point_header().padded_content_size().into() ;
+        let arc = triomphe::Arc::from_header_and_fn(
+            h,byte_size,
+            copy_from
+        );
+        let pkt= NetPktArc(arc);
         pkt.check(skip_hash)?;
         Ok(pkt)
     }
@@ -184,7 +188,7 @@ pub fn build() {
     assert_eq!(p2.point_parts, p3.point_parts);
     let a2 = boxed.as_netarc();
     use std::mem::size_of_val;
-    assert_eq!(size_of_val(&*a2.0), 32 + 32 + 4 + 5 + 3); // 3 is padding
+    assert_eq!(size_of_val(&*a2.0), 32 + 32 + 4 + 5 + 7); // 3 is padding
     assert_eq!(size_of_val(&*a2.0), size_of_val(&*a2.0));
 
     println!("Outer {:p}", arced.0.heap_ptr());