+++ ffi/linkspace-js/src/jspkt.rs
@@ -137,8 +137,9 @@ impl Pkt {
     }
 
     #[wasm_bindgen(getter)]
+    /// unpadded point size
     pub fn point_size(&self) -> u16{
-        self.0.point_header_ref().point_size.get()
+        self.0.point_header_ref().upoint_size()
     }
     #[wasm_bindgen(getter)]
     pub fn path_len(&self) -> Option<u8> {